+++ castor.php
@@ -1,6 +1,6 @@
 <?php
 
-\define('CASTOR_USE_CHDIR', true);
+\defined('CASTOR_USE_CHDIR') || \define('CASTOR_USE_CHDIR', true);
 
 use Castor\Attribute\AsTask;