+++ examples/advanced/castor-file/shebang.php
@@ -1,6 +1,8 @@
 #!/usr/bin/env -S castor --castor-file
 <?php
 
+\defined('CASTOR_USE_CHDIR') || \define('CASTOR_USE_CHDIR', true);
+
 use Castor\Attribute\AsTask;
 
 use function Castor\io;