+++ tests/Slow/RepackHelper.php
@@ -29,7 +29,7 @@ class RepackHelper
         $fs->dumpFile($castorAppDirPath . '/castor.php', <<<'PHP'
             <?php
 
-            define('CASTOR_USE_CHDIR', true);
+            defined('CASTOR_USE_CHDIR') || define('CASTOR_USE_CHDIR', true);
 
             use Castor\Attribute\AsTask;